DELETE License
Deletion for new license to be appended to the master branch.

NOTICE: Any copies that contain this deleted license (downloaded prior to deletion) are still under that license and do not follow the new license. However, any downloaded copies with the new license are subject to the new license to follow this one.